Add files via upload
Farooq Mp3 Player is an simple music player software which has 4 buttons in it and it plays .mp3 format files as it cannot recognize other type of music files. This project has its own speciality because the codes can be modified and can be updated even by the intermediate python programmers and make it more advance.